api|REFACT: adjust context package functions
